Applications have been localized (supporting english and croatian)
Applications (both desktop and mobile) are now localized, i.e. they are
now supported in english and in croatian language.
In desktop application user can manually select desired language and in
mobile application it is automatically selected one that is equal to
system's language or default one is being used.